Extract BuildableBackedProvider from BuildableBackedSetProvider
in preparation for a Provider<ResolvedComponentResult>
